requirements.txt
backoff==2.2.1
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.7
Deprecated==1.2.15
Flask==3.1.0
Flask-Cors==5.0.0
googleapis-common-protos==1.66.0
grpcio==1.68.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
opencensus-proto==0.1.0
opentelemetry-api==1.29.0
opentelemetry-distro==0.50b0
opentelemetry-exporter-opencensus==0.50b0
opentelemetry-exporter-otlp==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asyncio==0.50b0
opentelemetry-instrumentation-click==0.50b0
opentelemetry-instrumentation-dbapi==0.50b0
opentelemetry-instrumentation-flask==0.50b0
opentelemetry-instrumentation-grpc==0.50b0
opentelemetry-instrumentation-jinja2==0.50b0
opentelemetry-instrumentation-logging==0.50b0
opentelemetry-instrumentation-requests==0.50b0
opentelemetry-instrumentation-sqlite3==0.50b0
opentelemetry-instrumentation-threading==0.50b0
opentelemetry-instrumentation-urllib==0.50b0
opentelemetry-instrumentation-urllib3==0.50b0
opentelemetry-instrumentation-wsgi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
packaging==24.2
protobuf==3.20.3
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.0
zipp==3.21.0
